Assign Button & Worker Search Modal Implemented
The assign button now changes its text from either ‘Asignar’ or
‘Cambiar’ depending upon if a worker is assigned to the report.
The assign button in the modal also calls the server method if a worker
is selected.It shows an error if the worker is not selected.

The workerName is now shown on the details view after a worker is
selected.

The autocomplete fields are also updated after assigning a new worker
to a report.